implementato todo di arduino

+++ orchestrator/src/constants.rs
@@ -1,2 +1,2 @@
-pub const ARM_LENGTH: f64 = 151.1;
-pub const WATER_TIME: f64 = 10.0;
+pub const ARM_LENGTH: f32 = 151.1;
+pub const WATER_TIME: f32 = 10.0;
